use custom build script
the cloudflare vite plugin has no support for doing weird custom things
like combining the backend and frontend...at least with html. I tried
making multiple vite endpoints but it was impossible to retrieve the
built files. I assume they embed the vite metadata in the worker somehow
with the plugin, but I don't want to mess with that.